Move transformation methods to store
