+++ src/main/java/org/dita/dost/util/XMLUtils.java
@@ -20,7 +20,6 @@ import net.sf.saxon.serialize.MessageWarner;
 import net.sf.saxon.trans.XPathException;
 import net.sf.saxon.trans.XsltController;
 import org.apache.xml.resolver.tools.CatalogResolver;
-import org.dita.dost.exception.DITAOTException;
 import org.dita.dost.log.DITAOTLogger;
 import org.dita.dost.log.LoggingErrorListener;
 import org.dita.dost.module.saxon.DelegatingCollationUriResolver;
@@ -43,8 +42,6 @@ import java.util.stream.Stream;
 
 import static javax.xml.XMLConstants.DEFAULT_NS_PREFIX;
 import static javax.xml.XMLConstants.NULL_NS_URI;
-import static org.apache.commons.io.FileUtils.deleteQuietly;
-import static org.apache.commons.io.FileUtils.moveFile;
 import static org.dita.dost.util.Constants.*;
 
 /**
@@ -560,131 +557,6 @@ public final class XMLUtils {
             }
         }
         return buf.toString();
-    }
-
-    /**
-     * Transform file with XML filters. Only file URIs are supported.
-     *
-     * @param input absolute URI to transform and replace
-     * @param filters XML filters to transform file with, may be an empty list
-     */
-    public void transform(final URI input, final List<XMLFilter> filters) throws DITAOTException {
-        assert input.isAbsolute();
-        if (!input.getScheme().equals("file")) {
-            throw new IllegalArgumentException("Only file URI scheme supported: " + input);
-        }
-
-        transform(new File(input), filters);
-    }
-
-    /**
-     * Transform file with XML filters.
-     *
-     * @param inputFile file to transform and replace
-     * @param filters XML filters to transform file with, may be an empty list
-     */
-    public void transform(final File inputFile, final List<XMLFilter> filters) throws DITAOTException {
-        final File outputFile = new File(inputFile.getAbsolutePath() + FILE_EXTENSION_TEMP);
-        transformFile(inputFile, outputFile, filters);
-        try {
-            deleteQuietly(inputFile);
-            moveFile(outputFile, inputFile);
-        } catch (final RuntimeException e) {
-            throw e;
-        } catch (final Exception e) {
-            throw new DITAOTException("Failed to replace " + inputFile + ": " + e.getMessage());
-        }
-    }
-
-    /**
-     * Transform file with XML filters.
-     *
-     * @param inputFile input file
-     * @param outputFile output file
-     * @param filters XML filters to transform file with, may be an empty list
-     */
-    public void transform(final File inputFile, final File outputFile, final List<XMLFilter> filters) throws DITAOTException {
-        if (inputFile.equals(outputFile)) {
-            transform(inputFile, filters);
-        } else {
-            transformFile(inputFile, outputFile, filters);
-        }
-    }
-
-    private void transformFile(final File inputFile, final File outputFile, final List<XMLFilter> filters) throws DITAOTException {
-        if (!outputFile.getParentFile().exists() && !outputFile.getParentFile().mkdirs()) {
-            throw new DITAOTException("Failed to create output directory " + outputFile.getParentFile().getAbsolutePath());
-        }
-
-        try (final InputStream in = new BufferedInputStream(new FileInputStream(inputFile));
-             final OutputStream out = new BufferedOutputStream(new FileOutputStream(outputFile))) {
-            XMLReader reader = getXMLReader();
-            for (final XMLFilter filter : filters) {
-                // ContentHandler must be reset so e.g. Saxon 9.1 will reassign ContentHandler
-                // when reusing filter with multiple Transformers.
-                filter.setContentHandler(null);
-                filter.setParent(reader);
-                reader = filter;
-            }
-
-            final Serializer result = processor.newSerializer(out);
-            final ContentHandler serializer = result.getContentHandler();
-            reader.setContentHandler(serializer);
-
-            final InputSource inputSource = new InputSource(in);
-            inputSource.setSystemId(inputFile.toURI().toString());
-
-            reader.parse(inputSource);
-        } catch (final RuntimeException e) {
-            throw e;
-        } catch (final Exception e) {
-            throw new DITAOTException("Failed to transform " + inputFile + ": " + e.getMessage(), e);
-        }
-    }
-
-    /**
-     * Transform file with XML filters.
-     *
-     * @param input input file
-     * @param output output file
-     * @param filters XML filters to transform file with, may be an empty list
-     */
-    public void transform(final URI input, final URI output, final List<XMLFilter> filters) throws DITAOTException {
-        if (input.equals(output)) {
-            transform(input, filters);
-        } else {
-            transformURI(input, output, filters);
-        }
-    }
-
-    private void transformURI(final URI input, final URI output, final List<XMLFilter> filters) throws DITAOTException {
-        final File outputFile = new File(output);
-        if (!outputFile.getParentFile().exists() && !outputFile.getParentFile().mkdirs()) {
-            throw new DITAOTException("Failed to create output directory " + outputFile.getParentFile().getAbsolutePath());
-        }
-
-        try {
-            XMLReader reader = getXMLReader();
-            for (final XMLFilter filter : filters) {
-                // ContentHandler must be reset so e.g. Saxon 9.1 will reassign ContentHandler
-                // when reusing filter with multiple Transformers.
-                filter.setContentHandler(null);
-                filter.setParent(reader);
-                reader = filter;
-            }
-
-            final Serializer result = processor.newSerializer(outputFile);
-            final ContentHandler serializer = result.getContentHandler();
-            reader.setContentHandler(serializer);
-
-            final InputSource inputSource = new InputSource(input.toString());
-
-            reader.parse(inputSource);
-        } catch (final RuntimeException e) {
-            throw e;
-        } catch (final Exception e) {
-            throw new DITAOTException("Failed to transform " + input + ": " + e.getMessage(), e);
-        }
     }
 
     /** Close input source. */
